+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Add method `Attachable.unregister(HTMLElement)`
+
+### Fixed
+
+- Fix method `HasHTMLElement.style(String property, String value, boolean important)`
+
 ## [1.2.5] - 2023-10-24
 
 ### Added